split configuration.rs
